# Conflicts:
#	.idea/misc.xml

+++ src/entities/MovableEntity.java
@@ -12,7 +12,7 @@ public abstract class MovableEntity extends Entity {
     // בנאי: מקבל רק מיקום, מהירות ותמונה. (בלי גודל!)
     public MovableEntity(float x, float y, BufferedImage sprite) {
 
-        // קורא לבנאי של Entity שכתבת למעלה
+        // קורא לבנאי של Entity שכתבת למעלה ך
         // הוא מעביר לו רק את מה שהוא צריך: x, y, sprite
         super(x, y, sprite);
 
@@ -22,8 +22,8 @@ public abstract class MovableEntity extends Entity {
     }
 
     public void move() {
-        x += dx * speed* Time.deltaTime;
-        y += dy * speed* Time.deltaTime;
+        x += dx * speed* (float) Time.deltaTime;
+        y += dy * speed* (float)Time.deltaTime;
     }
 
     // פונקציות עזר לכיוונים
